Prefer forward conversion if registered
When a forward conversion is registered to getFreezeConversionMap(),
that explicit conversion is now preferred over the (implicit) inverse of
a (backward) conversion registered to getFreezeConversionInverseMap().
This behavior is consistent with the preexisting preference for explicit
forward conversion of certain simple types(_freeze_types_plain) and non-
iterables and strings (i.e., excluded by isIterableNotString()).